Move `cooperative_thread_hopper` to `runtime`
